pyproject.toml
[project]
name = "pollypm"
version = "0.1.0"
description = "Tmux-first orchestration for interactive CLI coding agents."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pyyaml>=6.0",
  "textual>=6.1.0",
  "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
]

[project.scripts]
pm = "pollypm.cli:app"
pollypm = "pollypm.cli:app"

[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pollypm = [
  "plugins_builtin/*/pollypm-plugin.toml",
  "defaults/rules/*.md",
  "defaults/magic/*.md",
  "defaults/docs/*.template",
  "defaults/docs/reference/*.md",
  "work/flows/*.yaml",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
